% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_one_subject.R
\name{process_one_subject}
\alias{process_one_subject}
\title{Stage I: Accelerometry data processing
Stage II: GPS data processing}
\usage{
process_one_subject(acc_file_path, gps_file_path)
}
\description{
Stage I: Accelerometry data processing
Stage II: GPS data processing
}
